requirements.txt
openai>=1.12.0
anthropic>=0.18.0
ollama>=0.1.0
gradio>=4.19.0
python-dotenv>=1.0.0